+++ Memento-iOS/Memento-iOS/Source/Presentation/Onboarding/View/SleepCycleSettingView.swift
@@ -11,15 +11,10 @@ import MDSKit
 struct SleepCycleSettingView: View {
     @State private var wakeUpTime: Date? = nil
     @State private var windDownTime: Date? = nil
-    @State private var isPickerPresented: Bool = false 
+    @State private var isPickerPresented: Bool = false
     @State private var selectedTimeType: TimeType = .wakeUp
-    @Binding var path: [String]
-
-    enum TimeType {
-        case wakeUp
-        case windDown
-    }
-
+    @Binding var path: [OnBoardingNavigationDestination]
+    
     var body: some View {
         ZStack {
             BackgroundView()
@@ -45,13 +40,13 @@ struct SleepCycleSettingView: View {
                 )
                 .padding(.horizontal, 16)
                 .padding(.top, 80)
-              
+                
                 Spacer()
                 
                 NextButton(wakeUpTime: $wakeUpTime, windDownTime: $windDownTime, path: $path)
                     .padding(.horizontal, 16)
                     .padding(.bottom, 10)
-              
+                
             }
         }
         .sheet(isPresented: $isPickerPresented) {
@@ -67,8 +62,8 @@ struct SleepCycleSettingView: View {
 
 // MARK: - CustomNavigationBar
 private struct CustomNavigationBar: View {
-    @Binding var path: [String]
-
+    @Binding var path: [OnBoardingNavigationDestination]
+    
     var body: some View {
         HStack(alignment: .top) {
             Button {
@@ -84,7 +79,7 @@ private struct CustomNavigationBar: View {
             Spacer()
             
             Button {
-                path.append("WorkSelectionView") // 다음 화면으로 이동
+                path.append(.workSelection) // 다음 화면으로 이동
             } label: {
                 Text("Skip")
                     .applyFont(.body_b_14)
@@ -125,7 +120,7 @@ private struct TimeSelectionView: View {
     @Binding var wakeUpTime: Date?
     @Binding var windDownTime: Date?
     @Binding var isPickerPresented: Bool
-    @Binding var selectedTimeType: SleepCycleSettingView.TimeType
+    @Binding var selectedTimeType: TimeType
     
     var body: some View {
         VStack(alignment: .leading, spacing: 29) {
@@ -167,7 +162,7 @@ private struct TimeSelectionView: View {
             Spacer()
             
             Button(action: action) {
-                Text(time.map { timeFormatter.string(from: $0) } ?? "00:00 AM")
+                Text(time.map { $0.formattedDate(with: "hh:mm a") } ?? "00:00 AM")
                     .applyFont(.body_r_14)
                     .foregroundColor(.white)
                     .frame(maxWidth: .infinity, alignment: .center)
@@ -175,6 +170,7 @@ private struct TimeSelectionView: View {
                     .background(Color.gray09)
             }
             .frame(width: 94, height: 36)
+
         }
     }
 }
@@ -183,12 +179,12 @@ private struct TimeSelectionView: View {
 private struct NextButton: View {
     @Binding var wakeUpTime: Date?
     @Binding var windDownTime: Date?
-    @Binding var path: [String]
-
+    @Binding var path: [OnBoardingNavigationDestination]
+    
     var body: some View {
         Button {
             if wakeUpTime != nil && windDownTime != nil {
-                path.append("WorkSelectionView")
+                path.append(.workSelection)
             }
         } label: {
             Text("Next")
@@ -198,61 +194,9 @@ private struct NextButton: View {
                 .frame(maxWidth: .infinity)
         }
         .cornerRadius(2)
-        .frame(height: 50) 
+        .frame(height: 50)
         .background((wakeUpTime != nil && windDownTime != nil) ? Color.green : Color.gray10)
     }
-}
-
-// MARK: - Time Picker View
-private struct TimePickerView: View {
-    @Binding var isPickerPresented: Bool
-    @Binding var selectedTimeType: SleepCycleSettingView.TimeType
-    @Binding var wakeUpTime: Date?
-    @Binding var windDownTime: Date?
-    
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            VStack {
-                Spacer()
-                
-                DatePicker(
-                    "",
-                    selection: Binding(
-                        get: {
-                            selectedTimeType == .wakeUp ? (wakeUpTime ?? Date()) : (windDownTime ?? Date())
-                        },
-                        set: { newValue in
-                            selectedTimeType == .wakeUp ? (wakeUpTime = newValue) : (windDownTime = newValue)
-                        }
-                    ),
-                    displayedComponents: .hourAndMinute
-                )
-                .datePickerStyle(WheelDatePickerStyle())
-                .colorScheme(.dark) // 다크 모드 강제 적용
-                .labelsHidden()
-            }
-            .presentationDetents([.fraction(0.3)]) // 시트 높이 설정
-            .presentationBackground(Color.gray09)
-            
-            // Done 버튼
-            Button(action: {
-                isPickerPresented = false
-            }) {
-                Text("Done")
-                    .foregroundColor(Color.white)
-                    .applyFont(.body_b_14)
-            }
-            .padding(.top, 14)
-            .padding(.trailing, 16)
-        }
-    }
-}
-
-// MARK: - Time Formatter
-private var timeFormatter: DateFormatter {
-    let formatter = DateFormatter()
-    formatter.dateFormat = "hh:mm a" // 00:00 PM 형식
-    return formatter
 }
 
 #Preview {
